requirements.txt
certifi==2022.9.24
charset-normalizer==2.1.1
gitdb==4.0.10
GitPython==3.1.29
idna==3.4
python-dotenv==0.21.0
requests==2.28.1
smmap==5.0.0
urllib3==1.26.13
